adapters in separate folders, fixed folders names, added class for the Scanner
