+++ adepter/src/main/java/Reader.java
@@ -1,0 +1,5 @@
+import java.util.Scanner;
+
+public class Reader {
+    public static Scanner scan = new Scanner(System.in);
+}
