+++ adepter/src/main/java/reportgenerator/ReportGenerator.java
@@ -1,4 +1,4 @@
-package reportGenerator;
+package reportgenerator;
 
 public interface ReportGenerator {
     void generateReport(String bodyReport);
