+++ adepter/src/main/java/reportgenerator/adapterexcel/ExcelReportAdapter.java
@@ -1,9 +1,11 @@
-package reportGenerator;
+package reportgenerator.adapterexcel;
 
-import reportGenerator.typeReport.ExcelReporter;
+import reportgenerator.ReportGenerator;
+import reportgenerator.typereport.ExcelReporter;
 
 public class ExcelReportAdapter implements ReportGenerator {
     ExcelReporter reporter = new ExcelReporter();
+
     @Override
     public void generateReport(String bodyReport) {
         reporter.createExcelFile(bodyReport);
