+++ adepter/src/main/java/reportgenerator/adapterpdf/PDFReportAdapter.java
@@ -1,0 +1,13 @@
+package reportgenerator.adapterpdf;
+
+import reportgenerator.ReportGenerator;
+import reportgenerator.typereport.PDFReporter;
+
+public class PDFReportAdapter implements ReportGenerator {
+    PDFReporter reporter = new PDFReporter();
+
+    @Override
+    public void generateReport(String bodyReport) {
+        reporter.createPDFFile(bodyReport);
+    }
+}
