+++ adepter/src/main/java/reportgenerator/typereport/PDFReporter.java
@@ -1,15 +1,12 @@
-package reportGenerator.typeReport;
-
-import java.io.DataInput;
-import java.io.FileNotFoundException;
-import java.io.FileOutputStream;
-import java.util.Date;
+package reportgenerator.typereport;
 
 import com.itextpdf.text.Document;
 import com.itextpdf.text.DocumentException;
 import com.itextpdf.text.Paragraph;
 import com.itextpdf.text.pdf.PdfWriter;
 
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
 
 
 public class PDFReporter {
@@ -22,7 +19,7 @@ public class PDFReporter {
             document.add(new Paragraph(content));
             document.close();
             writer.close();
-            System.out.println( "PDF отчет сформирован");
+            System.out.println("PDF отчет сформирован");
         } catch (DocumentException e) {
             e.printStackTrace();
         } catch (FileNotFoundException e) {
